requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
paddlepaddle>=2.4.0
paddleocr>=2.6.1
python-dotenv>=0.19.0
aiobotocore
pydantic-settings
groq